# Streaming contribution and fast contribution i/o