--- a/filecoin-proofs/Cargo.toml
+++ b/filecoin-proofs/Cargo.toml
@@ -48,7 +48,7 @@ sha2 = { version = "0.8.3", package = "sha2ni" }
 typenum = "1.11.2"
 bitintr = "0.3.0"
 gperftools = { version = "0.2", optional = true }
-phase2 = { package = "phase21", version = "0.6.0" }
+phase2 = { version = "0.7.0", package = "phase21" }
 simplelog = "0.8.0"
 rand_chacha = "0.2.1"
 dialoguer = "0.6.2"