@DrPeterVanNostrand 2020-06-26
Whoo, those are nice benchmarks.