@DrPeterVanNostrand 2020-06-28
Just a few comments, mostly about logs.