@porcuquine 2020-06-28
@DrPeterVanNostrand I think I have addressed all of your comments in my latest commit. In some cases, the resolution was different than your suggestion, for reasons I have clarified in my responses above.